+++ app/rust/src/frb_generated.rs
@@ -39,7 +39,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.12.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -600023791;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -752784205;
 
 // Section: executor
 
@@ -1548,6 +1548,58 @@ fn wire__crate__api__nullchat__UmbraApp_list_messages_impl(
                     &*api_that_guard,
                     api_contact_hex,
                     api_limit,
+                )?;
+                Ok(output_ok)
+            })())
+        },
+    )
+}
+fn wire__crate__api__nullchat__UmbraApp_merge_contact_impl(
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "UmbraApp_merge_contact",
+            port: None,
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_that = <RustOpaqueMoi<
+                flutter_rust_bridge::for_generated::RustAutoOpaqueInner<UmbraApp>,
+            >>::sse_decode(&mut deserializer);
+            let api_from_hex = <String>::sse_decode(&mut deserializer);
+            let api_into_hex = <String>::sse_decode(&mut deserializer);
+            deserializer.end();
+            transform_result_sse::<_, String>((move || {
+                let mut api_that_guard = None;
+                let decode_indices_ =
+                    flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![
+                        flutter_rust_bridge::for_generated::LockableOrderInfo::new(
+                            &api_that, 0, false,
+                        ),
+                    ]);
+                for i in decode_indices_ {
+                    match i {
+                        0 => api_that_guard = Some(api_that.lockable_decode_sync_ref()),
+                        _ => unreachable!(),
+                    }
+                }
+                let api_that_guard = api_that_guard.unwrap();
+                let output_ok = crate::api::nullchat::UmbraApp::merge_contact(
+                    &*api_that_guard,
+                    api_from_hex,
+                    api_into_hex,
                 )?;
                 Ok(output_ok)
             })())
@@ -3591,19 +3643,19 @@ fn pde_ffi_dispatcher_primary_impl(
         24 => {
             wire__crate__api__nullchat__UmbraApp_gif_search_impl(port, ptr, rust_vec_len, data_len)
         }
-        40 => wire__crate__api__nullchat__UmbraApp_read_attachment_impl(
+        41 => wire__crate__api__nullchat__UmbraApp_read_attachment_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        58 => wire__crate__api__nullchat__UmbraApp_start_network_impl(
+        59 => wire__crate__api__nullchat__UmbraApp_start_network_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        63 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
+        64 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }
@@ -3667,68 +3719,69 @@ fn pde_ffi_dispatcher_sync_impl(
         ),
         30 => wire__crate__api__nullchat__UmbraApp_list_groups_impl(ptr, rust_vec_len, data_len),
         31 => wire__crate__api__nullchat__UmbraApp_list_messages_impl(ptr, rust_vec_len, data_len),
-        32 => wire__crate__api__nullchat__UmbraApp_messages_from_contact_impl(
+        32 => wire__crate__api__nullchat__UmbraApp_merge_contact_impl(ptr, rust_vec_len, data_len),
+        33 => wire__crate__api__nullchat__UmbraApp_messages_from_contact_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        33 => wire__crate__api__nullchat__UmbraApp_my_invite_impl(ptr, rust_vec_len, data_len),
-        34 => wire__crate__api__nullchat__UmbraApp_my_onion_impl(ptr, rust_vec_len, data_len),
-        35 => wire__crate__api__nullchat__UmbraApp_my_picture_impl(ptr, rust_vec_len, data_len),
-        36 => wire__crate__api__nullchat__UmbraApp_open_impl(ptr, rust_vec_len, data_len),
-        37 => wire__crate__api__nullchat__UmbraApp_open_account_impl(ptr, rust_vec_len, data_len),
-        38 => {
+        34 => wire__crate__api__nullchat__UmbraApp_my_invite_impl(ptr, rust_vec_len, data_len),
+        35 => wire__crate__api__nullchat__UmbraApp_my_onion_impl(ptr, rust_vec_len, data_len),
+        36 => wire__crate__api__nullchat__UmbraApp_my_picture_impl(ptr, rust_vec_len, data_len),
+        37 => wire__crate__api__nullchat__UmbraApp_open_impl(ptr, rust_vec_len, data_len),
+        38 => wire__crate__api__nullchat__UmbraApp_open_account_impl(ptr, rust_vec_len, data_len),
+        39 => {
             wire__crate__api__nullchat__UmbraApp_open_account_auto_impl(ptr, rust_vec_len, data_len)
         }
-        39 => {
+        40 => {
             wire__crate__api__nullchat__UmbraApp_pending_messages_impl(ptr, rust_vec_len, data_len)
         }
-        41 => wire__crate__api__nullchat__UmbraApp_rename_contact_impl(ptr, rust_vec_len, data_len),
-        42 => wire__crate__api__nullchat__UmbraApp_rename_group_impl(ptr, rust_vec_len, data_len),
-        43 => wire__crate__api__nullchat__UmbraApp_repair_tor_impl(ptr, rust_vec_len, data_len),
-        44 => wire__crate__api__nullchat__UmbraApp_safety_number_impl(ptr, rust_vec_len, data_len),
-        45 => {
+        42 => wire__crate__api__nullchat__UmbraApp_rename_contact_impl(ptr, rust_vec_len, data_len),
+        43 => wire__crate__api__nullchat__UmbraApp_rename_group_impl(ptr, rust_vec_len, data_len),
+        44 => wire__crate__api__nullchat__UmbraApp_repair_tor_impl(ptr, rust_vec_len, data_len),
+        45 => wire__crate__api__nullchat__UmbraApp_safety_number_impl(ptr, rust_vec_len, data_len),
+        46 => {
             wire__crate__api__nullchat__UmbraApp_search_messages_impl(ptr, rust_vec_len, data_len)
         }
-        46 => wire__crate__api__nullchat__UmbraApp_send_file_impl(ptr, rust_vec_len, data_len),
-        47 => wire__crate__api__nullchat__UmbraApp_send_gif_impl(ptr, rust_vec_len, data_len),
-        48 => wire__crate__api__nullchat__UmbraApp_send_group_message_impl(
+        47 => wire__crate__api__nullchat__UmbraApp_send_file_impl(ptr, rust_vec_len, data_len),
+        48 => wire__crate__api__nullchat__UmbraApp_send_gif_impl(ptr, rust_vec_len, data_len),
+        49 => wire__crate__api__nullchat__UmbraApp_send_group_message_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        49 => {
+        50 => {
             wire__crate__api__nullchat__UmbraApp_send_over_network_impl(ptr, rust_vec_len, data_len)
         }
-        50 => wire__crate__api__nullchat__UmbraApp_set_autologin_impl(ptr, rust_vec_len, data_len),
-        51 => {
+        51 => wire__crate__api__nullchat__UmbraApp_set_autologin_impl(ptr, rust_vec_len, data_len),
+        52 => {
             wire__crate__api__nullchat__UmbraApp_set_contact_saved_impl(ptr, rust_vec_len, data_len)
         }
-        52 => wire__crate__api__nullchat__UmbraApp_set_contact_status_impl(
+        53 => wire__crate__api__nullchat__UmbraApp_set_contact_status_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        53 => wire__crate__api__nullchat__UmbraApp_set_custom_bridges_impl(
+        54 => wire__crate__api__nullchat__UmbraApp_set_custom_bridges_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        54 => wire__crate__api__nullchat__UmbraApp_set_duress_passphrase_impl(
+        55 => wire__crate__api__nullchat__UmbraApp_set_duress_passphrase_impl(
             ptr,
             rust_vec_len,
             data_len,
         ),
-        55 => wire__crate__api__nullchat__UmbraApp_set_gif_key_impl(ptr, rust_vec_len, data_len),
-        56 => wire__crate__api__nullchat__UmbraApp_set_my_picture_impl(ptr, rust_vec_len, data_len),
-        57 => wire__crate__api__nullchat__UmbraApp_set_verified_impl(ptr, rust_vec_len, data_len),
-        59 => wire__crate__api__nullchat__UmbraApp_user_code_impl(ptr, rust_vec_len, data_len),
-        60 => wire__crate__api__nullchat__UmbraApp_username_impl(ptr, rust_vec_len, data_len),
-        61 => wire__crate__api__nullchat__app_version_impl(ptr, rust_vec_len, data_len),
-        62 => wire__crate__api__simple__greet_impl(ptr, rust_vec_len, data_len),
-        64 => wire__crate__api__nullchat__install_update_impl(ptr, rust_vec_len, data_len),
-        65 => wire__crate__api__nullchat__offered_update_impl(ptr, rust_vec_len, data_len),
-        66 => wire__crate__api__nullchat__set_native_dir_impl(ptr, rust_vec_len, data_len),
+        56 => wire__crate__api__nullchat__UmbraApp_set_gif_key_impl(ptr, rust_vec_len, data_len),
+        57 => wire__crate__api__nullchat__UmbraApp_set_my_picture_impl(ptr, rust_vec_len, data_len),
+        58 => wire__crate__api__nullchat__UmbraApp_set_verified_impl(ptr, rust_vec_len, data_len),
+        60 => wire__crate__api__nullchat__UmbraApp_user_code_impl(ptr, rust_vec_len, data_len),
+        61 => wire__crate__api__nullchat__UmbraApp_username_impl(ptr, rust_vec_len, data_len),
+        62 => wire__crate__api__nullchat__app_version_impl(ptr, rust_vec_len, data_len),
+        63 => wire__crate__api__simple__greet_impl(ptr, rust_vec_len, data_len),
+        65 => wire__crate__api__nullchat__install_update_impl(ptr, rust_vec_len, data_len),
+        66 => wire__crate__api__nullchat__offered_update_impl(ptr, rust_vec_len, data_len),
+        67 => wire__crate__api__nullchat__set_native_dir_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }